requirements.txt
apache-airflow==1.10.1
slackclient==1.3.0